Create a DOM element for the task when the task is constructed.
Instead of having to call task.createTaskDOMElement() on every task,
it seems to make more sense to create the DOM elements on Task
construction and then save them as a property of the Task object.
This may be a problem if there are a ton of tasks though? I'm
really not sure.