Don't add 'active-task' to the task.element by default.
I've changed up the logic of how the Task DOM Element is created.
Before, I was actually calling a function to do this, but now every
Task has a DOM element generated on Task construction. Because of this,
I noticed I was doing something dumb by adding 'active-task' to
every task.element on creation, then having to strip this class
in different places where it wasn't needed. It makes way more sense
to only add this class when it is needed.