Change createListInput 'keydown' -> 'keyup' eventListener.
When using the 'keydown' event, the event listener would run, but
if there was only 1 character in the input, it wouldn't be
accessible in the this.value property. This meant that the save
button was still disabled when the list had a name with only 1
character. When using the 'keyup' event, the value will be present
and the logic works as expected.