Add comment to clarify last portion of centerCol 'click' event handler.
I don't have the final portion of the centerCol 'click' event handler
scoped inside a conditional, I just assume that by that point all
other options have been filtered out and the user is attempting to
edit a Task. This seems like a fine assumption at the moment, but
it may be smarter to have this inside some sort of guard statement.