Associate DOM elements generated server-side with the correct Task objects in the taskList array.
Whenever the browser makes a server-request for a page (like on
page reload), I'm generating the DOM elements server-side, and then
loading in all the Task and TaskList information and building the
client-side objects for each. Each Task and TaskList will have a
DOM element associated with it when it is created, but I wasn't
taking into account the fact that the references for the TaskList
loaded from the server weren't being setup correctly. Now, when
the page loads, the currently loaded TaskList's Tasks have their
'element' properties assigned to the DOM element representing that
Task, instead of just an arbitrary DOM element created by the
constructor.