Add updateTaskListUI() call to handler for creating a new TaskList.
I pulled the wiping of the DOM out of updatePageState() and moved
it to updateTaskListUI() in an earlier commit, and apparently
when a new TaskList is created I was relying on updatePageState()
to wipe the UI, and not calling updateTaskListUI(). Originally,
updateTaskListUI() was just populating the TaskListUI from the
tasks in the taskList, but for a new taskList there would be no
reason to run this. Now, I have to run it.