I messed up and cowboy coded a ton of stuff without committing, see the description for details :<(
OK, I got carried away because I'm tired and was doing too many things
and coded a ton of stuff without committing. Here is a breakdown:

1. Added a completedTaskHeader property and method to create it to
the TaskList constructor
2. Added the siteIcon to the TaskList constructor
3. Refactored the 'click' listener attached to the leftCol that
runs when the taskListNav elements are clicked to use the new Task
List methods.

It was really irresponsible to code like that, but I'm wore out and
just wanted to get the app in a working state tonight.