Remove some old commented-out code and a REFACTOR? note.
When a new task is created in the createTaskInput 'keydown' event
handler, I would originally query the centerCol for any list elements
and if there were any, I would remove the site-icon. This seems silly,
as soon as a Task is added, there will be tasks, so the icon should
be removed.