Change logic in TaskList.prototype.addTask to only deal with the TaskList, not DOM operations.
I was performing the logic to add a new Task to the TaskLists tasks
array, and then make update calls to /taskLists and /tasks, and THEN
performing DOM manipulation. This effectively cancels-out my whole
reason for not doing page reloads, since I was having to wait for
the Task creation call in the DB to finish so I could get the _id
property of the Task so TaskList.addTask() would work correctly. Now,
addTask() simply adds the Task to the tasks array and calls the
updateTaskListDB() function. The DOM manipulation can happen right
away when a new Task is created, and all the DB logic can work
itself out after.