Remove assignments to task.taskList.
I'm not using the taskList property of a Task. I think I had some
reason to set this, but for the life of me can't figure out why
at this point. Since I have activeTaskList as a global variable,
the Task itself doesn't need a reference to the TaskList it is
a part of, but maybe that was why I originally set this?