Remove TaskList.prototype.getDBProps().
I needed this before because I had a circular referenece in the
Task object because the Task.taskList property pointed to a TaskList
that had the current Task as part of it's tasks array. I knew something
was fishy about my design, but was too tired to fix it yesterday so
I made this method as a patch till my brain started working again.