Ensure that updateTaskListUI() has the activeTask as the first param and doesn't assign to activeTask.
I've got some bad logic here, this smells. This commit begins to
fix it. updateTaskListUI takes in the taskListToAdd and taskListToRemove,
but where I was calling it sometimes taskListToAdd was activeTaskList
and sometimes taskListToRemove was activeTaskList. I was also assigning
activeTaskList to taskListToAdd at the end of the method. Now, whenever
I call updateTaskListUI I'm always passing in activeTaskList as the
first param.