Refactor changePageURL() to use global activeTaskList variable.
I know globals are supposed to be bad, but it just makes more sense
to me to have this activeTaskList global variable and to use it
for updating the UI and page state. The alternative is to query
the DOM each time to figure out which taskList is loaded and then
filter the taskLists array for that TaskList, which seems silly.