Refactor updateTaskListUI to reference global activeTaskList.
I was passing the activeTaskList in as the first param, but this
is not neccessary since I have the global activeTaskList variable
to work with. I know globals are bad, but having a few around makes
this logic so much easier and keeps me from having to query all the
taskLists each time I need to work with the active one.