Create TaskContainer.prototype.add() and refactor all calls to taskContainer.prepend to use this function.
I want to toggle the siteIcon and completedTaskToggle whenever
there is an item added to the TaskContainer, so I moved that logic
into the TaskContainer.prototype.add() function. This should save
me a lot of trouble down the line when I will be deleting things
from taskContainers and will have to be toggling the siteIcon
and completedTaskToggle.