Refactor updateTaskListUI to take in newActiveTaskList as a param.
I'm changing the way I think about this method, but not what it does.
The naming convention I'm using now reflects the fact that the
current activeTaskList is being wiped from the DOM, while the
newActiveTaskList is being added to the DOM. I don't reassign
activeTaskList before calling this method, I want to wait until
all operations are completed, as this makes more sense to me.