Move logic for toggling completedTaskList display into completedTaskToggle 'click' event handler.
I'm not liking the big global event handlers on the main columns
as much, so I'm pulling out the major sections and attaching
handlers to them to make things more readable and easier to
grasp at first glance. I may go back to the larger event-delgation
model at the end of the project, but going forward having scoped
event listeners seems easier to deal with and understand.