Setup a conditional around the section of the centerCol 'click' event handler related to taskList editing.
I realized that if I'm going to attach a high-level event handler
such as the centerCol 'click' handler, I'm going to have to have
every section that runs scoped inside an if() conditional, or I
will have to have a bunch of guard conditions to filter out all
unwanted clicks from hitting the logic not scoped to an if(). Not
sure why I didn't see this earlier, but the code will be much more
readable and secure now.