In centerCol 'click' handler section that handles task editing, remove logic to select the task's span element.
When the user clicks on a Task in one of the taskContainers, I
wanted the text to have focus so they could edit the taskName. I
thought I needed to call focus() on the span element of the task
that was clicked, but it turns out it doesn't matter if the li or
the span element is clicked, the focus() method behaves the same.