Change conditional for firing .active-task toggle.
I had the logic all wrong for this section of the centerCol 'click'
event handler. What I thought I was doing was adding focus to the
clicked element, the setting it to be the .active-task. Apparently,
since the span has the contenteditable attribute, the browser is
handling the focus part for me, so that was not needed. So, all
I really need to do here is toggle the .active-class.