Refactor logic in centerCol 'click' handler that sets the .active-task to the clicked task.
I was doing some silly things, but now I'm not. Originally I was
checking to see if there was an element with .active-task, then removing that
class from that element, but there will always be an .active-task
when a task element has been clicked, as anytime there are task elements
in the TaskContainer there is an active one, so this check was
not needed.