Change the logic of taskListNavContainer 'click' event handler newActiveTaskList selection.
Before, I was using an if..else determined by the tagName of the
event target to determine which version of find() to run on the
taskLists array in order to find the newActiveTaskList. Now, I just
have an extra variable, taskListLink, that will hold the anchor tag
and I can use this to have only one version of find() which searches
through the taskLists array for the taskList with a url matching the
href attribute of the taskListLink.