Make if() conditional to run toggleCompletedStatus() more specific.
In the centerCol 'click' event handler, where I want to listen for
the click's to <i> elements to toggle the completed status of a
Task, I was calling toggleCompletedStatus for any <i> element
that was clicked. This threw a bug when the completedTaskToggle
caret icon was clicked though, so I had to add a check for
that icon.