Change the placement of completedTaskList nav setup in changeActiveTaskList().
It matters where the logic to setup the navigation between the
completedTaskList appears in changeActiveTaskList(). It has to
come after updateTaskListUI(), since that calls populateTaskContainers()
which updates the completedTaskToggle, which I want to override
when the completedTaskList is displayed. This seems like something
that will need to be refactored down the line for sure.