Setup correct way to add a Task to a TaskList.
I was using some strange logic before to add a Task to a TaskList.
For some reason, I would send the entire task to PUT /taskLists,
and then find the TaskList by the ownerId property of the Task, and
add the _id of the Task to the TaskList. This was dumb. Now, I
send just the task._id to POST /taskLists/:taskListId/tasks.