Implement DELETE /taskLists/:id/tasks route-handler.
I ran into an issue here that I haven't encountered yet. When comparing
the task from the the taskList.tasks array pulled from the DB with
the task._id passed in at req.body._id, simple string comparison
did not work. The task represents an ObjectID, so I needed to call
the ObjectId.equals() method to compare it with another ObjectID
or a string representing the ObjectID.