Don't allow toggleCompletedStatus to run if activeTaskList is Trash.
When the user is on the Trash page, they should not be able to toggle
the completed status of a Task. I may go so far as to remove the toggle
icon eventually, but I'm trying to keep this project as a strict clone
of TickTick and that developer does not remove the icon, rather just
makes it un-usable.