Ensure a taskList is the activeTaskList before triggering siteIcon after removing a Task.
In TaskList.prototype.removeTask(), I remove the Task from the
TaskList.tasks array client-side and in the DB. I then trigger
the siteIcon based on whether or not there are any tasks left in
the taskList.tasks array. There was a subtle bug hiding here though,
as if the activeTaskList was not the taskList that removeTask() was
called on, the siteIcon may display when there are still tasks on
the screen. The specific case was when I had the 'Inbox' taskList
as the activeTaskList, and then had one incomplete and one complete
task. If I deleted the complete task, and there were no other
completed tasks in the completedTaskList, the siteIcon would
display. I solved the issue by adding a check to ensure that the
siteIcon is only displayed if the taskList is the activeTaskList
AND the taskList.tasks array is empty.