Add some whitespace to ejs logic to hide ellipsis in dashboard.ejs.
I don't like the way I am structuring my ejs logic to show/hide
the ellipsis, but I can't figure out a better way at this point.
Currently, there is an if..else that is really only toggling a
single class in a span element, but I'm using it to toggle the
entire element instead to keep it readable. Having the if..else
logic stuck inside of an HTML elements class attribute is just
unreadable.