Fixed bug in toggleDumpsterIcon().
For some reason, I was grabbing every <span> element in the
pageTitleContainer and toggling the .hidden class on them, instead
of just grabbing the dumpsterIcon element and toggling the hidden
class on it. This lead to a hard to track down bug where the
ellipsis icon was being toggled by this method. Why I wrote it this
way in the first place is a mystery to me, I must've been completely
out of it.