Don't set '.active-task' on page load.
I don't think that the first task in the incomplete task container
should be automatically set as the '.active-task' when the page
loads, or when a new task list is navigated too. Originally, this
seemed like a good idea, but now I only want to set the '.active-task'
class on a task if it has just been added or if it has just been
clicked to edit.