Refine query used to select taskListNavElements in setTaskListNavElements.
I made a change to the taskListNavElements, where I gave the
addListButton a '.task-list-nav-item' class so it could have the
nav item styling. In setTaskListNavElements(), I was querying
the '.task-list-nav-item' selector, but I only wanted the
taskListNavElements that were NOT the addListButton, so I adjusted
the selector using the ':not()' pseudo selector.