Ensure only current date gets '.current-date'.
I had a bug where I was determining which <td> element got the
'.current-date' class. I was only using the Date.prototype.getDate()
function to compare the day of the month with the current date, but
I also needed to make sure to check that the month was the current
month.