Overload the hideDialogBox() function for calendarDialogBox.
When the hideDialogBox() function is called on the calendarDialogBox,
I need to rebuild the calendar so that it shows the current month.
That way, if a user closes the calendar but they have navigated to
a different month, the current month will still be displayed if they
open the calendar again.